more refactoring and moving things around
