+++ _test/_loadTestConfig.js
@@ -1,10 +1,10 @@
 const path = require('path')
-const locateProjectRoot = require('../utilities/locateProjectRoot')
-const configLoader = require('../config/loader')
+const locateProjectRoot = require('../src/utilities/locateProjectRoot')
+const configLoader = require('../src/config/loader')
 
 function loadTestConfig () {
   const projectRoot = locateProjectRoot()
-  const dirParts = [projectRoot, 'src', '_test', '_testConfig', 'config.js']
+  const dirParts = [projectRoot, '_test', '_testConfig', 'config.js']
   const filePath = dirParts.join(path.sep)
   return configLoader.load(filePath)
 }
