+++ src/AppServer.spec.js
@@ -1,5 +1,5 @@
-const AppServer = require('../AppServer')
-const Endpoint = require('../Endpoint')
+const AppServer = require('./AppServer')
+const Endpoint = require('./Endpoint')
 const Request = require('supertest')
 
 describe('AppServer', () => {
